+++ scrapyd_k8s/joblogs/log_handler_k8s.py
@@ -226,7 +226,6 @@ class KubernetesJobLogHandler:
         -------
         None
         """
-        self.object_storage_provider = LibcloudObjectStorage(self.config)
         try:
 
             pod = event['object']
